Fix "implicit conversion from float to int" in captcha
The params are even prefixed with an `i`, yet they are floats.
The x and y parameters of imagefttext need to be ints.

```
PHP Unknown Error(8192): Implicit conversion from float 10.5 to int loses precision in core/inc/3p/Class_Captcha.php line 320

Backtrace starts in www/index.php line 32.
GWF3->__construct()......... gwf3.class.php line 107.
GWF3->init()................ gwf3.class.php line 173.
GWF3->onLoadModule()........ gwf3.class.php line 435.
GWF_Module->execute()....... core/inc/util/GWF_Module.php line 342.
GWF_Captcha->execute()...... core/module/GWF/method/Captcha.php line 38.
PhpCaptcha->Create()........ core/inc/3p/Class_Captcha.php line 398.
PhpCaptcha->DrawCharacters() core/inc/3p/Class_Captcha.php line 320.
imagefttext()............... [unknown file] line ?.
GWF_Debug::error_handler().. core/inc/util/GWF_Debug.php line 180.
```

+++ core/inc/3p/Class_Captcha.php
@@ -312,9 +312,9 @@
             $aCharDetails = imageftbbox($iFontSize, $iAngle, $sCurrentFont, $this->sCode[$i], array());
             
             // calculate character starting coordinates
-            $iX = $this->iSpacing / 4 + $i * $this->iSpacing;
+            $iX = (int) ($this->iSpacing / 4 + $i * $this->iSpacing);
             $iCharHeight = $aCharDetails[2] - $aCharDetails[5];
-            $iY = $this->iHeight / 2 + $iCharHeight / 4; 
+            $iY = (int) ($this->iHeight / 2 + $iCharHeight / 4);
             
             // write text to image
             imagefttext($this->oImage, $iFontSize, $iAngle, $iX, $iY, $iTextColour, $sCurrentFont, $this->sCode[$i], array());
